ON-611: Add breadcrumbs to station detail page
To reduce duplication of code on multiple pages and page types, a breadcrumbs partial component was created and replaces that functionality that was used in the article and gallery pages.

Since a good deal of shared functionality exists, it was consolidated within the breadcrumbs component in index.js. Which must be imported into pages that intend to use this functionality. Its basic function is to automatically create links based on properties provided.

Another reason that this was done this way is that the render method is not available to non-instance components, or more generally, any work it does is ignored.

Since this is basically a "dumb" component, it was decided not to make this into an instance component.