ON-478: Bug Fix - Video not displaying correctly
Fixes include:

Client: checking data-is-playlist value against the string 'true', rather than just checking for truthiness, as 'false' as a string is truthy in JS

Model: Not 100% certain why this is needed and once we are no longer importing, it probably won't be, but for now, there are times when videoId is not being overridden by contentId in upgrade.js for imports, and this seems to cause problems when attempting to edit

Upgrade: Needed to set contentId, not content from videoId. Also, videoSource was not defaulting to video on imported items. Wrote code so that it could handle if imports actually have playlists for some reason.