NOJIRA: expose 'clay.radio.com' via docker network
This fixes a bug where during local development, 'clay.radio.com'
wasn't resolving to the nginx container like it should.

Had the nginx container not depended on (linked to) the clay container
then this would have been solved by linking the clay container to
nginx.  Circular links are not possible though, so the solution per

https://medium.com/@yani/two-way-link-with-docker-compose-8e774887be41

is to create a network and declare an alias which enables all networked
containers to access the machine via the aliased host name.