ON-103: add explicit option 'shouldDedupeContent'
Because searchByQuery is called by methods which sometimes shouldn't
be deduped but still need to pass locals, we can no longer assume
the existence of locals means the content should be deduped.  I should
have written the code like this to begin with  anyway because it
both makes the code more explicit and prevents issues in the future
where someone passes locals for another reason and doesn't realize
doing so causes content to be deduped.

also
- remove an unused file universal/latest-content.js which is a
  duplicate of universal/recirc-cmpt.js