ON-892: revert the work that's left
This branch added logic to prevent publishing of articles and galleries
from the back and frontends.  The backend checks were removed when
the station access permissions were enforced.  The frontend kiln
hooks are still here though and needed to be removed.

Note this means this ticket is essentially removed from development,
save for we still need to test the station access for articles and
galleries.  We're still working on getting that testing relocated
to another ticket.