ON-1144: add pg dependency
note the version is meant to be in sync with the one
amphora-storage-postgres declares.

this is necessary due to a bug in amphora-storage-postgres.  If npm
decides amphora-storage-postgres needs its own version of pg due to
package conflicts but knex can be top-level, then knex doesn't have
access to pg and throws an error when attempting to initialize
the driver.

amphora-storage-postgres should instead be declaring pg as a peer
dependency so the app owners are responsible for having pg
installed, where it will be installed at the top level giving knex
access to it.