ON-837 | csd | Fix for noIndexNoFollow
Also discovered during fix that imports that publish via the import do
not automatically add the `url`, `published` and `publishTime` values to
the page meta values. For this reason, the query was modified to use
pages with ids ending in `@published` to check if it was published, and
it uses the `customUrl` from page data if the `url` does not exist it
also uses the latest `updateTime`, in place of `publishTime` if
`publishTime` does not exist.

It now verifies that the page is an article by only joining the article
component table.