ON-754: update image component
copy the functionality added to 'feed-image' into the 'image' component.
Ideally I think 'image' would have been used for 'feed-image' as well
but I couldn't think of a way to have the schema publish the image url
depending on other instance data.

to be explicit, the functionality added was to assign image meta data
to the instance so we could warn editors when they upload an image that
doesn't fit msn's requirements.

also add styling for the warnings in both image and feed-image