ON-754: add msnTitle to article
since the msn feed requires a title of at least 20 characters, we need a
separate field for it.  And since I couldn't figure out a way to have a good UX
for a field which conditionally overrides another, this commit adds a new kiln
input 'default-text-with-override' which provides a good ux.