ON-1333: update existing content
when we release the smart news feed we want existing content to appear in the
feed.  The only reasonable way to do that at the moment is to re-publish
existing content so it gets updated to be compatible with the feed.

this commit also updates the feed to not check for "The Associated Press" byline
since content with that will always be marked "no index / no follow"