ON-1336: update existing content
mostly copied from ON-1333 commit 587ddec

when we release the msn news feed we want existing content to appear
in the feed.  The only reasonable way to do that at the moment is to
re-publish existing content so it gets updated to be compatible with
the feed.

this commit also updates the feed to not check for "The Associated
Press" byline since content with that will always be marked "no index
/ no follow"