ON-1499: Fix nav dropdowns on marketing pages.
There was a problem in some components that would gather data from elements on the page,
though the data was not being properly defaulted if those elements were not found.
This was causing the navigation dropdowns not to show up, as the errors from the above
would cause the main javascript to stop.