ON-1524: Improve top-nav functionality and cleanliness
The setMobileNavOnResize function was set to trigger a menu toggle on a window resize event.
Originally, this function was meant to prevent an open navigation menu from persisting
when resizing a window from a mobile size to a desktop size. However, due to how
mobile browsers show/hide the browser navigation bar when scrolling, the window size
would change, triggering the resize event. This caused the mobile navigation menu to
constantly open and close as the user scrolled up and down on the page.

To resolve the issue, I adjusted the function to only close navigation when resizing,
instead of toggling. To achieve this, I separated the functions that open and close
navigation into explicit "open" and "close" functions, which the "toggle" functions
will dynamically choose from based on current state. Additionally, I separated the
functions that open the mobile navigation from those that open the desktop navigation,
and wait until the user has finished resizing the window before triggering the close functions.