compute divider class in model.js
The CSS class was getting a bit complicated,
but beyond that, for some reason @root.locals.sectionFront
would reference the wrong value when navigating from one
section front to another using the top nav. By computing
this value in model.js, the sectionFront is always correct.
As an added bonus, the template file is now easier to read.