ON-2105: fix invalid token error
after updating the auth data we need to check if the auth.token exists
since for some reason it's not being persisted in sync with the user.
Ideally we'd instead attach the auth information directly with the
user object on the session, however then we'd have to implement some
cleanup since amphora puts the user object directly onto locals which
is sent to the frontend which would be insecure.