ON-1422 store podcasts sitemap in xml in db
Change the name of the route - it was getting caught by an express middleware
searching for anything including "podcasts" in the route.
Use new utils from ON-1637 - mainly, the createUrl() function. This will ensure we
always generate the appropriate URL, even if we change the format.
Disable redis cache on getPodcastsFromApi - caching the results was increasing
the compute time tenfold. This was causing a 504 error when the
/update-db-podcast-data route was called, as it would take over a minute to resolve.
Add loading dots to the migration, so users know that the program hasn't hung
as it waits for the server to update the db podcast data.
Improve overall cleanliness of code - in some cases, also efficiency.