ON-2398: make station condition a should clause
previously this was an array which was assumed to be a 'should' clause.
addMust didn't work since it would declare that content had to be both
original content to the station as well as syndicated to the station.
The solution is to make it a should, which applies to both addShould
and addMust condition types.